requirements.txt
streamlit>=1.28.0
torch>=2.0.0
transformers>=4.30.0
sentence-transformers>=2.2.2
numpy>=1.24.0
pandas>=2.0.0
pyyaml>=6.0
